Revert "App refactor"
