+++ Isle of Trust/src/App.tsx
@@ -3,16 +3,20 @@ import * as RK from "react-konva";
 import "./css/App.css";
 import Konva from "konva";
 import useImage from "use-image";
+import * as util from "./utilities";
+import {
+    AnimResources,
+    AnimInfluence,
+    AnimChoice,
+    AnimMood,
+    AnimChangeIdeology,
+} from "./models/animation";
 
 import { Face, Hat, GeneratePawn } from "./generators/pawn";
-import { PlayerSidebar } from "./components/playerSideBar";
-import { SelectedSidebar } from "./components/selectedSideBar";
-import { Board } from "./components/board";
 import { Grid } from "./generators/map";
 import {
-    MetaAgent,
     Agent,
-    DeadAgent,
+    AGENT_RADIUS,
     Relation,
     Ideology,
     Personality,
@@ -31,6 +35,10 @@ import {
 import { isAccordionItemSelected } from "react-bootstrap/esm/AccordionContext";
 */
 import { KonvaEventObject } from "konva/lib/Node";
+import { getActiveElement } from "@testing-library/user-event/dist/utils";
+import { ThemeConsumer } from "react-bootstrap/esm/ThemeProvider";
+import { timingSafeEqual } from "crypto";
+import { allowedNodeEnvironmentFlags } from "process";
 /*
 import { timeStamp } from "console";
 */
@@ -53,28 +61,30 @@ export const MAP_URL: { [key: string]: string } = {
     Ring: "url(../Maps/mapRing.png)",
     Spokes: "url(../Maps/mapSpokes.png)",
     Crescent: "url(../Maps/mapCrescent.png)",
+    Small: "url(../Maps/mapSmall.png)",
 };
+
 //export let MAP_INDEX = 0;
 let currentMap = "Pronged";
 
 const DIFFICULTY_VALUES: { [key: string]: number } = {
-    easy: 20,
+    easy: 19,
     medium: 15,
     hard: 10,
     extreme: 5,
 };
 
 class SidebarState {
-    player: MetaAgent;
-    selected: MetaAgent;
+    player: Agent;
+    selected: Agent;
     playerToSelected: Relation | undefined;
     selectedToPlayer: Relation | undefined;
     influenceChoices: SpendingContainer;
 
     constructor(
-        map: Graph<MetaAgent, Relation>,
-        player: MetaAgent,
-        selected: MetaAgent
+        map: Graph<Agent, Relation>,
+        player: Agent,
+        selected: Agent
     ) {
         this.player = player;
         this.selected = selected;
@@ -85,14 +95,16 @@ class SidebarState {
 }
 
 interface GameViewState {
-    map: Graph<MetaAgent, Relation>;
+    map: Graph<Agent, Relation>;
     sidebarState: SidebarState;
-    select: (agent: MetaAgent) => void;
+    select: (agent: Agent) => void;
     turnCount: number;
+    selectCharacterDisplay: boolean;
 }
 
 export interface StartInfo {
     //Using strings until it's connected up
+    name: string;
     hat: string;
     face: string;
     ideologyColor: string;
@@ -120,9 +132,12 @@ class GameView extends React.Component<StartInfo, GameViewState> {
             map.getVertices()[
                 Math.floor(Math.random() * map.getVertices().length)
             ];
+        
+        //generates player with chosen face/hat/name/ideology
         if (player instanceof Agent) {
             player.face = Face[props.face as keyof typeof Face];
             player.hat = Hat[props.hat as keyof typeof Hat];
+            player.name = props.name
 
             switch (props.ideologyColor) {
                 case "9ec4ea":
@@ -160,7 +175,7 @@ class GameView extends React.Component<StartInfo, GameViewState> {
         let selected = map.getVertices()[0];
         let sidebarState = new SidebarState(map, player, selected);
 
-        let select = (agent: MetaAgent) => {
+        let select = (agent: Agent) => {
             sidebarState.selected = agent;
             sidebarState.playerToSelected = map.getEdge(player, agent)!;
             sidebarState.selectedToPlayer = map.getEdge(agent, player)!;
@@ -173,14 +188,19 @@ class GameView extends React.Component<StartInfo, GameViewState> {
             sidebarState: sidebarState,
             select: select,
             turnCount: turnCount,
+            selectCharacterDisplay: false,
         };
+
+        //Needed for setState function
+        this.deselectCharacter = this.deselectCharacter.bind(this);
+
         //checking to see if props are coming in
         console.log("GameView");
         console.log(props);
     }
 
     tallyChoicesForAllNeighbors(
-        map: Graph<MetaAgent, Relation>,
+        map: Graph<Agent, Relation>,
         you: Agent
     ): choiceTally {
         const neighbors = map.getEdges(you);
@@ -197,7 +217,7 @@ class GameView extends React.Component<StartInfo, GameViewState> {
         return sumChoices;
     }
 
-    countTotalInfluence(map: Graph<MetaAgent, Relation>, agent: Agent): String {
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String {
         const neighbors = map.getEdges(agent);
         let totalInfluence = 0;
         let numberOfNeighbors = 0;
@@ -235,36 +255,30 @@ class GameView extends React.Component<StartInfo, GameViewState> {
         this.generateRound(edges);
         this.drainResources(vertices);
 
-        this.handleDead(vertices);
         this.forceUpdate();
     }
 
-    drainResources(vertices: MetaAgent[]) {
+    drainResources(vertices: Agent[]) {
         vertices.forEach((v1) => {
-            if (v1 instanceof Agent) {
-                v1.resources -= RESOURCE_LOST_PER_TURN;
-            }
+            v1.resources -= RESOURCE_LOST_PER_TURN;
         });
     }
 
-    drainInfluence(edges: [MetaAgent, MetaAgent, Relation][]) {
+    drainInfluence(edges: [Agent, Agent, Relation][]) {
         edges.forEach(([v1, v2, e]) => {
-            if (v2 instanceof Agent) {
-                const v2Agent = v2 as Agent;
-                const maxInfluenceChange =
-                    BASE_INFLUENCE_LOST_PER_TURN *
-                    v2Agent.getInfluenceability();
-                e.influence = e.incrementAttributeBy(
-                    -maxInfluenceChange,
-                    e.influence
-                );
-            }
+            const v2Agent = v2 as Agent;
+            const maxInfluenceChange =
+                BASE_INFLUENCE_LOST_PER_TURN *
+                v2Agent.getInfluenceability();
+            e.influence = e.incrementAttributeBy(
+                -maxInfluenceChange,
+                e.influence
+            );
         });
     }
 
-    handleInfluenceChanges(vertices: MetaAgent[]) {
+    handleInfluenceChanges(vertices: Agent[]) {
         vertices.forEach((v1) => {
-            if (v1 instanceof Agent) {
                 const v1Relations = this.state.map.getEdges(v1)!;
                 let spendingMap = new SpendingContainer();
                 if (v1 === this.state.sidebarState.player) {
@@ -283,121 +297,58 @@ class GameView extends React.Component<StartInfo, GameViewState> {
                         );
                 });
                 this.driftIdeology(v1);
-            }
         });
     }
 
-    generateRound(edges: [MetaAgent, MetaAgent, Relation][]) {
+    generateRound(edges: [Agent, Agent, Relation][]) {
         const turnsToSample: number = 10;
         edges.forEach(([v1, v2, e1]) => {
-            if (v1 instanceof Agent) {
-                if (v2 instanceof Agent) {
-                    const e2 = this.state.map.getEdge(v2, v1);
-                    if (v1.id < v2.id && e2 instanceof Relation) {
-                        const v1Strat = v1.ideology.toStrategy();
-                        const v2Strat = v2.ideology.toStrategy();
-                        const v1Choice = generateChoice(
-                            v1Strat,
-                            v1.mood,
-                            e2.history
-                        );
-                        const v2Choice = generateChoice(
-                            v2Strat,
-                            v2.mood,
-                            e1.history
-                        );
+            const e2 = this.state.map.getEdge(v2, v1);
+            if (v1.id < v2.id && e2 instanceof Relation) {
+                const v1Strat = v1.ideology.toStrategy();
+                const v2Strat = v2.ideology.toStrategy();
+                const v1Choice = generateChoice(
+                    v1Strat,
+                    v1.mood,
+                    e2.history
+                );
+                const v2Choice = generateChoice(
+                    v2Strat,
+                    v2.mood,
+                    e1.history
+                );
 
-                        let resourceChange = v1.resources;
-                        let moodChange = v1.mood;
+                let resourceChange = v1.resources;
+                let moodChange = v1.mood;
 
-                        v1.rewardResources(v1Choice, v2Choice);
-                        v2.rewardResources(v1Choice, v2Choice);
-                        v1.updateMood(v1Choice, v2Choice);
-                        v2.updateMood(v1Choice, v2Choice);
+                v1.rewardResources(v1Choice, v2Choice);
+                v2.rewardResources(v1Choice, v2Choice);
+                v1.updateMood(v1Choice, v2Choice);
+                v2.updateMood(v1Choice, v2Choice);
 
-                        resourceChange = v1.resources - resourceChange;
-                        moodChange = v1.mood - moodChange;
+                resourceChange = v1.resources - resourceChange;
+                moodChange = v1.mood - moodChange;
 
-                        e1.history.addTurn(new Turn(v1Choice, v1Choice));
-                        e2.history.addTurn(new Turn(v2Choice, v2Choice));
+                e1.history.addTurn(new Turn(v1Choice, v1Choice));
+                e2.history.addTurn(new Turn(v2Choice, v2Choice));
 
-                        let opinionChange = e1.opinion;
-                        e1.updateOpinion(
-                            e2.influence,
-                            e2.history.getAvgChoice(turnsToSample),
-                            v1.personality.getVolatility()
-                        );
-                        e2.updateOpinion(
-                            e1.influence,
-                            e1.history.getAvgChoice(turnsToSample),
-                            v2.personality.getVolatility()
-                        );
-                        opinionChange = e1.opinion - opinionChange;
-                    }
-                }
+                let opinionChange = e1.opinion;
+                e1.updateOpinion(
+                    e2.influence,
+                    e2.history.getAvgChoice(turnsToSample),
+                    v1.personality.getVolatility()
+                );
+                e2.updateOpinion(
+                    e1.influence,
+                    e1.history.getAvgChoice(turnsToSample),
+                    v2.personality.getVolatility()
+                );
+                opinionChange = e1.opinion - opinionChange;
             }
         });
 
         this.setState((state) => {
             return { turnCount: this.state.turnCount + 1 };
-        });
-    }
-
-    handleDead(vertices: MetaAgent[]) {
-        vertices.forEach((v1) => {
-            // If agent has recently hit 0 resources, replace with dead agent. Keep relations but
-            // set influence to 0. Update graph with dead agent and modified relations
-            if (v1 instanceof Agent && v1.resources < 0) {
-                const vDead = new DeadAgent(v1.coords, v1.id);
-                let v1Edges = this.state.map.getEdges(v1);
-                if (v1Edges) {
-                    v1Edges.forEach((e1, v2) => {
-                        e1.influence = 0;
-                        this.state.map.getEdges(v2)?.forEach((e2, v3) => {
-                            if (v3 === v1) {
-                                this.state.map.getEdges(v2)?.delete(v1);
-                                this.state.map.getEdges(v2)?.set(vDead, e2);
-                            }
-                        });
-                    });
-
-                    this.state.map.updateVertex(v1, vDead, v1Edges);
-                }
-            }
-            // If dead agent has been dead long enough, generate new agent with modified
-            // relations and update graph.
-            else if (v1 instanceof DeadAgent && v1.deadCount === 3) {
-                // Make new random agent but keep meta data.
-                const vNew = new Agent(
-                    v1.id.toString(),
-                    new Ideology(10, 10),
-                    new Personality(10, 10),
-                    // TODO: these should be lower
-                    100,
-                    100,
-                    v1.id,
-                    v1.coords
-                );
-                // Maintain edge history but reset influence for new agent
-                let vEdges = this.state.map.getEdges(v1);
-                if (vEdges) {
-                    vEdges.forEach((e1, v2) => {
-                        e1.influence = 0;
-                        this.state.map.getEdges(v2)?.forEach((e2, v3) => {
-                            if (v3 === v1) {
-                                this.state.map.getEdges(v2)?.delete(v1);
-                                this.state.map.getEdges(v2)?.set(vNew, e2);
-                            }
-                        });
-                    });
-
-                    this.state.map.updateVertex(v1, vNew, vEdges);
-                }
-            }
-            // If dead agent has not met maximum dead count, increment dead count
-            else if (v1 instanceof DeadAgent && v1.deadCount < 3) {
-                v1.deadCount += 1;
-            }
         });
     }
 
@@ -412,15 +363,15 @@ class GameView extends React.Component<StartInfo, GameViewState> {
         const ideologyAppeal: Map<Ideology, number> = new Map();
         const neighbors = this.state.map.getEdges(agent)!;
         let totalInfluence = 0;
-        neighbors.forEach((relation: Relation, neighbor: MetaAgent) => {
-            if (neighbor instanceof Agent) {
-                const theirInfluence = this.state.map.getEdge(
-                    neighbor,
-                    agent
-                )!.influence;
-                ideologyAppeal.set(neighbor.ideology, theirInfluence);
-                totalInfluence += theirInfluence;
-            }
+        
+        //calcautes the total influnce of all nehboors
+        neighbors.forEach((relation: Relation, neighbor: Agent) => {
+            const theirInfluence = this.state.map.getEdge(
+                neighbor,
+                agent
+            )!.influence;
+            ideologyAppeal.set(neighbor.ideology, theirInfluence);
+            totalInfluence += theirInfluence;
         });
 
         const drifts = new DriftContainer();
@@ -433,31 +384,476 @@ class GameView extends React.Component<StartInfo, GameViewState> {
         agent.driftIdeology(drifts);
     }
 
+    deselectCharacter(value: boolean) {
+        this.setState({selectCharacterDisplay: value});
+    }
+
+    render() {
+        //if there is a selected player display right sidebar
+        if (this.state.selectCharacterDisplay) {
+            return (
+                <div className="game">
+                        <Board
+                            map={this.state.map}
+                            turnCount={this.state.turnCount}
+                            selected={this.state.sidebarState.selected}
+                            select={this.state.select.bind(this)}
+                            player = {this.state.sidebarState.player}
+                            deselectCharacter={this.deselectCharacter}
+                        />
+                        <PlayerSidebar
+                            map={this.state.map}
+                            round={this.tempTurn.bind(this)}
+                            sidebarState={this.state.sidebarState}
+                            tallyChoicesNeighbors={this.tallyChoicesForAllNeighbors}
+                            countTotalInfluence={this.countTotalInfluence}
+                        />
+                        <SelectedSidebar
+                            map={this.state.map}
+                            round={this.tempTurn.bind(this)}
+                            sidebarState={this.state.sidebarState}
+                            tallyChoicesNeighbors={this.tallyChoicesForAllNeighbors}
+                            countTotalInfluence={this.countTotalInfluence}
+                            deselectCharacter={this.deselectCharacter}
+                        />
+                </div>
+            );
+        } else {
+            return (
+                <div className="game">
+                        <Board
+                            map={this.state.map}
+                            turnCount={this.state.turnCount}
+                            selected={this.state.sidebarState.selected}
+                            select={this.state.select.bind(this)}
+                            player = {this.state.sidebarState.player}
+                            deselectCharacter={this.deselectCharacter}
+                            
+                        />
+                        <PlayerSidebar
+                            map={this.state.map}
+                            round={this.tempTurn.bind(this)}
+                            sidebarState={this.state.sidebarState}
+                            tallyChoicesNeighbors={this.tallyChoicesForAllNeighbors}
+                            countTotalInfluence={this.countTotalInfluence}
+                        />
+                </div>
+            );
+        }
+
+    }
+}
+
+interface PlayerSidebarProps {
+    sidebarState: SidebarState;
+    map: Graph<Agent, Relation>;
+    tallyChoicesNeighbors: (
+        map: Graph<Agent, Relation>,
+        agent: Agent
+    ) => choiceTally;
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String;
+    round: () => void;
+}
+
+class PlayerSidebar extends React.Component<PlayerSidebarProps, unknown> {
     render() {
         return (
-            <div className="game">
-                <Board
-                    map={this.state.map}
-                    turnCount={this.state.turnCount}
-                    selected={this.state.sidebarState.selected}
-                    select={this.state.select.bind(this)}
-                    player = {this.state.sidebarState.player}
-                    current = {currentMap}
+            <div className="sidebar playerSidebar">
+                <PlayerDisplay
+                    map={this.props.map}
+                    sidebarState={this.props.sidebarState}
+                    tallyChoicesNeighbors={this.props.tallyChoicesNeighbors}
+                    countTotalInfluence={this.props.countTotalInfluence}
                 />
-                <PlayerSidebar
-                    map={this.state.map}
-                    round={this.tempTurn.bind(this)}
-                    sidebarState={this.state.sidebarState}
-                    tallyChoicesNeighbors={this.tallyChoicesForAllNeighbors}
-                    countTotalInfluence={this.countTotalInfluence}
+                <InfluenceMenu
+                    round={this.props.round}
+                    sidebarState={this.props.sidebarState}
+                    map={this.props.map}
                 />
-                <SelectedSidebar
-                    map={this.state.map}
-                    round={this.tempTurn.bind(this)}
-                    sidebarState={this.state.sidebarState}
-                    tallyChoicesNeighbors={this.tallyChoicesForAllNeighbors}
-                    countTotalInfluence={this.countTotalInfluence}
+            </div>
+        );
+    }
+}
+
+interface PlayerDisplayProps {
+    sidebarState: SidebarState;
+    map: Graph<Agent, Relation>;
+    tallyChoicesNeighbors: (
+        map: Graph<Agent, Relation>,
+        agent: Agent
+    ) => choiceTally;
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String;
+}
+
+class PlayerDisplay extends React.Component<PlayerDisplayProps> {
+    render() {
+        let choices = new choiceTally();
+        let name: string = "";
+        if (this.props.sidebarState.player instanceof Agent) {
+            const you = this.props.sidebarState.player as Agent;
+            choices = this.props.tallyChoicesNeighbors(this.props.map, you);
+            name = you.name;
+        }
+        return (
+            <div className="player-display">
+                <div className="agent-type">
+                    Player: <span className="agent-name">{name}</span>                 
+                </div>
+                <Display
+                    map={this.props.map}
+                    agent={this.props.sidebarState.player}
+                    agentChoices={choices}
+                    countTotalInfluence={this.props.countTotalInfluence}
                 />
+            </div>
+        );
+    }
+}
+
+interface InfluenceMenuProps {
+    round: () => void;
+    sidebarState: SidebarState;
+    map: Graph<Agent, Relation>;
+}
+
+class InfluenceMenu extends React.Component<InfluenceMenuProps> {
+    public spendingMap = new SpendingContainer();
+
+    render() {
+        const neighbors = this.props.map.getEdges(
+            this.props.sidebarState.player
+        )!;
+
+        if (this.props.sidebarState.player instanceof Agent) {
+            return (
+                <div className="influence-menu">
+                    <div className="influence-title">
+                        spend resources,
+                        <br /> influence your neighbors:
+                    </div>
+                    <InfluenceOptions
+                        selected={this.props.sidebarState.selected}
+                        player={this.props.sidebarState.player as Agent}
+                        neighbors={neighbors}
+                        spendingMap={this.props.sidebarState.influenceChoices}
+                    />
+                    <button onClick={this.props.round}>Confirm Choices</button>
+                </div>
+            );
+        }
+    }
+}
+
+interface InfluenceOptionsProps {
+    selected: Agent;
+    player: Agent;
+    neighbors: Map<Agent, Relation>;
+    spendingMap: SpendingContainer;
+}
+
+class InfluenceOptions extends React.Component<InfluenceOptionsProps> {
+    // this is an array of React components with a long and complicated type signature
+    private children: JSX.Element[] = [];
+    private resourcesGiveable: number;
+
+    constructor(props: any) {
+        super(props);
+
+        this.resourcesGiveable = this.props.player.resources;
+        this.allowResources = this.allowResources.bind(this);
+        this.redoChildren();
+    }
+
+    // a terrible workaround to force resourcesGiveable to update for all children every round
+    private redoChildren() {
+        this.children = [];
+        let key = 0;
+        if (this.props.neighbors) {
+            for (const entry of this.props.neighbors.entries()) {
+                const newChild = (
+                    <InfluenceEntry
+                        key={key}
+                        allowResources={this.allowResources}
+                        resourcesGiveable={this.props.player.resources}
+                        agent={entry[0]}
+                        spendingMap={this.props.spendingMap}
+                    />
+                );
+                this.children.push(newChild);
+                key += 1;
+            }
+        }
+    }
+
+    componentDidUpdate(
+        prevProps: Readonly<InfluenceOptionsProps>,
+        prevState: Readonly<{}>,
+        snapshot?: any
+    ): void {
+        this.resourcesGiveable = this.props.player.resources;
+    }
+
+    // Called by `InfluenceEntry`s.
+    // An entry may call this function to ask to increase its `initial` number of resources by `increment`.
+    // The function considers this and returns the new number of resources the entry can have.
+    allowResources(initial: number, increment: number): number {
+        const total = increment + initial;
+        // trying to give what you don't have
+        if (
+            (increment > 0 && total > this.props.player.resources) ||
+            (increment > 0 && this.resourcesGiveable <= 0)
+        ) {
+            return initial;
+        }
+        // trying to give a negative number
+        else if (
+            (increment < 0 &&
+                this.resourcesGiveable >= this.props.player.resources) ||
+            (increment < 0 && initial <= 0)
+        ) {
+            return 0;
+        } else {
+            this.resourcesGiveable -= increment;
+            return total;
+        }
+    }
+
+    render() {
+        this.redoChildren();
+        return <div className="influence-options">{this.children}</div>;
+    }
+}
+
+interface InfluenceEntryProps {
+    allowResources: (giving: number, increment: number) => number;
+    resourcesGiveable: number;
+    agent: Agent;
+    spendingMap: SpendingContainer;
+}
+
+interface InfluenceEntryState {
+    given: number;
+}
+
+class InfluenceEntry extends React.Component<
+    InfluenceEntryProps,
+    InfluenceEntryState
+> {
+    private containerRef = React.createRef<HTMLDivElement>();
+    private stageRef = React.createRef<Konva.Stage>();
+    private agentImageScale: number = 0.1;
+    private canvasWidth = AGENT_IMAGE_WIDTH * this.agentImageScale;
+    private canvasHeight = AGENT_IMAGE_HEIGHT * this.agentImageScale;
+
+    constructor(props: any) {
+        super(props);
+        this.state = { given: 0 };
+    }
+
+    updateGiven(increment: number) {
+        if (this.props.agent instanceof Agent) {
+            const newGiven = this.props.allowResources(
+                this.state.given,
+                increment
+            );
+            this.setState({ given: newGiven });
+            this.props.spendingMap.data.set(this.props.agent, newGiven);
+        }
+    }
+
+    render() {
+        const sMaybe = this.state.given === 1 ? "" : "s";
+        const givenString = String(this.state.given) + " resource" + sMaybe;
+        return (
+            <div className="influence-entry" ref={this.containerRef}>
+                <div className="influence-agent">
+                    <RK.Stage
+                        ref={this.stageRef}
+                        width={this.canvasWidth}
+                        height={this.canvasHeight}
+                    >
+                        <RK.Layer>
+                            <SidebarAgentImage
+                                canvasWidth={this.canvasWidth}
+                                data={this.props.agent}
+                            />
+                        </RK.Layer>
+                    </RK.Stage>
+                </div>
+                <div className="sidebar-agent-info">
+                    <button
+                        className="more-resource"
+                        onClick={() => {
+                            this.updateGiven(1);
+                        }}
+                    >
+                        +
+                    </button>
+                    <button
+                        className="less-resource"
+                        onClick={() => {
+                            this.updateGiven(-1);
+                        }}
+                    >
+                        -
+                    </button>
+                    <div>{givenString}</div>
+                </div>
+            </div>
+        );
+    }
+}
+
+interface SelectedSidebarProps {
+    sidebarState: SidebarState;
+    map: Graph<Agent, Relation>;
+    tallyChoicesNeighbors: (
+        map: Graph<Agent, Relation>,
+        agent: Agent
+    ) => choiceTally;
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String;
+    round: () => void;
+    deselectCharacter:(value: boolean) => void;
+}
+
+class SelectedSidebar extends React.Component<SelectedSidebarProps, unknown> {
+    render() {
+        return (
+            <div className="sidebar selectedSidebar">
+                <SelectedDisplay
+                    map={this.props.map}
+                    sidebarState={this.props.sidebarState}
+                    tallyChoicesNeighbors={this.props.tallyChoicesNeighbors}
+                    countTotalInfluence={this.props.countTotalInfluence}
+                    deselectCharacter={this.props.deselectCharacter}
+                />
+                <Stats
+                    sidebarState={this.props.sidebarState}
+                    tallyChoicesNeighbors={this.props.tallyChoicesNeighbors}
+                />
+            </div>
+        );
+    }
+}
+
+interface SelectedDisplayProps {
+    sidebarState: SidebarState;
+    map: Graph<Agent, Relation>;
+    tallyChoicesNeighbors: (
+        map: Graph<Agent, Relation>,
+        agent: Agent
+    ) => choiceTally;
+    
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String;
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String;
+    deselectCharacter:(value: boolean) => void;
+}
+
+class SelectedDisplay extends React.Component<SelectedDisplayProps> {
+    deselectCharacter(value: boolean) {
+        this.props.deselectCharacter(false);
+    }
+    render() {
+        let choices = new choiceTally();
+        let name = "";
+        if (this.props.sidebarState.selected instanceof Agent) {
+            const them = this.props.sidebarState.selected as Agent;
+            choices = this.props.tallyChoicesNeighbors(this.props.map, them);
+            name = them.name;
+        }
+        return (
+            <div className="selected-display">
+                <div className="agent-type" id="selected-character">
+                    <div>
+                        selected: <span className="agent-name">{name}</span>
+                    </div>
+                    <div 
+                        className="deselect"
+                        onClick={() => {
+                            this.deselectCharacter(false);
+                        }}>
+                        &#9746;
+                    </div> 
+                </div>
+                <Display
+                    map={this.props.map}
+                    agent={this.props.sidebarState.selected}
+                    agentChoices={choices}
+                    countTotalInfluence={this.props.countTotalInfluence}
+                />
+                <Judgement sidebarState={this.props.sidebarState} />
+            </div>
+        );
+    }
+}
+
+interface JudgementProps {
+    sidebarState: SidebarState;
+}
+
+class Judgement extends React.Component<JudgementProps> {
+    render() {
+        let judgement: string = "doesn't know you";
+        if (
+            this.props.sidebarState.player === this.props.sidebarState.selected
+        ) {
+            judgement = "that's you!";
+        } else if (
+            this.props.sidebarState.selectedToPlayer instanceof Relation &&
+            this.props.sidebarState.playerToSelected instanceof Relation
+        ) {
+            judgement =
+                this.props.sidebarState.selectedToPlayer.getDescriptiveOpinion() +
+                "; " +
+                this.props.sidebarState.playerToSelected.getDescriptiveInfluence();
+        }
+
+        return <div className="judgement">{judgement}</div>;
+    }
+}
+
+interface StatsProps {
+    sidebarState: SidebarState;
+    tallyChoicesNeighbors: (
+        map: Graph<Agent, Relation>,
+        agent: Agent
+    ) => choiceTally;
+}
+class Stats extends React.Component<StatsProps, unknown> {
+    render() {
+        const theirChoices = new choiceTally();
+        const yourChoices = new choiceTally();
+        let theySpent = 0;
+        let youSpent = 0;
+        if (this.props.sidebarState.selectedToPlayer) {
+            theirChoices.tallyChoices(
+                this.props.sidebarState.selectedToPlayer.history!
+            );
+            theySpent = this.props.sidebarState.selectedToPlayer.resourcesSpent;
+        }
+        if (this.props.sidebarState.playerToSelected) {
+            yourChoices.tallyChoices(
+                this.props.sidebarState.playerToSelected.history!
+            );
+            youSpent = this.props.sidebarState.playerToSelected.resourcesSpent;
+        }
+        return (
+            <div className="stats-container">
+                <div className="stats">
+                    <p>
+                        They've spent {theySpent} resources trying to influence
+                        you, while you've spent {youSpent} resources trying to
+                        influence them.
+                    </p>
+                    <p>
+                        They've given to you {theirChoices.gave} times, while
+                        you've given to them {yourChoices.gave} times.
+                    </p>
+                    <p>
+                        They've cheated you {theirChoices.cheated} times, while
+                        you've cheated them {yourChoices.cheated} times.
+                    </p>
+                </div>
             </div>
         );
     }
@@ -468,13 +864,13 @@ interface DisplayState {
 }
 
 interface DisplayProps {
-    map: Graph<MetaAgent, Relation>;
-    agent: MetaAgent;
+    map: Graph<Agent, Relation>;
+    agent: Agent;
     agentChoices: choiceTally;
-    countTotalInfluence(map: Graph<MetaAgent, Relation>, agent: Agent): String;
+    countTotalInfluence(map: Graph<Agent, Relation>, agent: Agent): String;
 }
 
-export class Display extends React.Component<DisplayProps, DisplayState> {
+class Display extends React.Component<DisplayProps, DisplayState> {
     private containerRef = React.createRef<HTMLDivElement>();
     private stageRef = React.createRef<Konva.Stage>();
     private resizeTimeout?: NodeJS.Timeout;
@@ -543,7 +939,6 @@ export class Display extends React.Component<DisplayProps, DisplayState> {
                         <RK.Layer>
                             <SidebarAgentImage
                                 canvasWidth={this.currentCanvasWidth}
-                                alive={this.props.agent.isAlive()}
                                 data={this.props.agent}
                             />
                         </RK.Layer>
@@ -571,7 +966,7 @@ export class Display extends React.Component<DisplayProps, DisplayState> {
 }
 
 interface MoodProps {
-    agent: MetaAgent;
+    agent: Agent;
 }
 
 function Mood(props: MoodProps) {
@@ -580,17 +975,12 @@ function Mood(props: MoodProps) {
     let moodDesc: String;
     let moodImgPath: string;
 
-    if (props.agent instanceof Agent) {
-        const agent = props.agent as Agent;
-        moodDesc = agent.getMoodDescription();
-        moodImgPath =
-            "images/mood-" +
-            moodDesc.split(" ").join("-").toLowerCase() +
-            ".png";
-    } else {
-        moodDesc = "dead";
-        moodImgPath = "images/mood-dead.png";
-    }
+    const agent = props.agent as Agent;
+    moodDesc = agent.getMoodDescription();
+    moodImgPath =
+        "images/mood-" +
+        moodDesc.split(" ").join("-").toLowerCase() +
+        ".png";
 
     return (
         <div className="mood">
@@ -605,15 +995,15 @@ function Mood(props: MoodProps) {
     );
 }
 
-interface AliveAgentImageProps {
+interface AgentImageProps {
     x: number;
     y: number;
     selected: boolean;
     data: Agent;
-    player: MetaAgent;
+    player: Agent;
 }
 
-export function AliveAgentImage(props: AliveAgentImageProps) {
+function AgentImage(props: AgentImageProps) {
     let imageNode = React.useRef<any>(null);
 
     let defaultScale = 0.12;
@@ -716,41 +1106,12 @@ export function AliveAgentImage(props: AliveAgentImageProps) {
     );
 }
 
-interface DeadAgentImageProps {
-    x: number;
-    y: number;
-}
-
-export function DeadAgentImage(props: DeadAgentImageProps) {
-    let imageNode: any = null;
-    const [deadImage] = useImage("pawns/gravestone.png");
-
-    const scale = 0.12;
-
-    return (
-        <RK.Image
-            ref={(node) => {
-                imageNode = node;
-            }}
-            image={deadImage}
-            x={props.x - (AGENT_IMAGE_WIDTH / 2) * scale}
-            y={props.y - (AGENT_IMAGE_HEIGHT / 2) * scale - 20}
-            scaleX={scale}
-            scaleY={scale}
-            offsetX={130 * scale * 2}
-            offsetY={145 * scale * 2}
-        />
-    );
-}
-
 interface SidebarAgentImageType {
     canvasWidth: number;
-    alive: boolean;
-    data: MetaAgent;
+    data: Agent;
 }
 
-export function SidebarAgentImage(props: SidebarAgentImageType) {
-    const [deadImage] = useImage("pawns/gravestone.png");
+function SidebarAgentImage(props: SidebarAgentImageType) {
     const scale = props.canvasWidth / AGENT_IMAGE_WIDTH;
 
     let face = Face.Glasses;
@@ -785,17 +1146,240 @@ export function SidebarAgentImage(props: SidebarAgentImageType) {
         }
     }
 
-    if (props.alive) {
-        return (
-            <RK.Group scaleX={scale} scaleY={scale}>
-                {GeneratePawn(hat, face, ideology)}
-            </RK.Group>
-        );
-    } else {
-        return <RK.Image image={deadImage} scaleX={scale} scaleY={scale} />;
-    }
+    return (
+        <RK.Group scaleX={scale} scaleY={scale}>
+            {GeneratePawn(hat, face, ideology)}
+        </RK.Group>
+    );
 }
 
+interface BoardProps {
+    map: Graph<Agent, Relation>;
+    turnCount: number;
+    selected: Agent;
+    select: (agent: Agent) => void;
+    player: Agent;
+    deselectCharacter: (value: boolean) => void;
+}
+
+class Board extends React.Component<BoardProps> {
+    private containerRef = React.createRef<HTMLDivElement>();
+    private stageRef = React.createRef<Konva.Stage>();
+
+    private resizeTimeout?: NodeJS.Timeout;
+
+    componentDidMount() {
+        this.resizeEvent = this.resizeEvent.bind(this);
+        this.resizeEvent();
+        window.addEventListener("resize", this.resizeEvent);
+    }
+
+    componentWillUnmount() {
+        window.removeEventListener("resize", this.resizeEvent);
+    }
+
+    resizeEvent() {
+        if (this.resizeTimeout) clearTimeout(this.resizeTimeout);
+
+        this.resizeTimeout = setTimeout(() => {
+            if (
+                this.containerRef === undefined ||
+                this.containerRef.current == null ||
+                this.stageRef.current == null
+            )
+                return;
+
+            const container = this.containerRef.current;
+            const stage = this.stageRef.current;
+
+            let scale = container.offsetWidth / SCENE_WIDTH;
+
+            stage.width(SCENE_WIDTH * scale);
+            stage.height(SCENE_HEIGHT * scale);
+            stage.scale({ x: scale, y: scale });
+        }, RESIZE_TIMEOUT);
+    }
+
+    select(pawn: Agent) {
+        this.props.select(pawn);
+    }
+
+    deselectCharacter(value: boolean) {
+        this.props.deselectCharacter(value)
+    }
+
+    render() {
+        return (
+            <div className="board">
+                <div
+                    className="map"
+                    ref={this.containerRef}
+                    style={{
+                        backgroundImage: MAP_URL[currentMap],
+                        backgroundSize: "100%",
+                    }}
+                >
+                    {/* <img src={require("./assets/Maps/prongedMap.png")} /> //not working  */}
+
+                    <RK.Stage ref={this.stageRef}>
+                        <RK.Layer>
+                            {this.props.map.getAllEdges().map(([v1, v2, e]) => (
+                                <RK.Line
+                                    key={v1.id + "->" + v2.id}
+                                    points={util.edgeLinePoints(
+                                        v1.coords[0],
+                                        v1.coords[1],
+                                        v2.coords[0],
+                                        v2.coords[1]
+                                    )}
+                                    strokeLinearGradientStartPointX={
+                                        v1.coords[0]
+                                    }
+                                    strokeLinearGradientStartPointY={
+                                        v1.coords[1]
+                                    }
+                                    strokeLinearGradientEndPointX={v2.coords[0]}
+                                    strokeLinearGradientEndPointY={v2.coords[1]}
+                                    strokeLinearGradientColorStops={[
+                                        0,
+                                        "white",
+                                        1,
+                                        util.colorFromWeight(e.influence),
+                                    ]}
+                                    opacity={0.6}
+                                    strokeWidth={4}
+                                    fillAfterStrokeEnabled={true}
+                                    pointerLength={AGENT_RADIUS / 3}
+                                    pointerWidth={AGENT_RADIUS / 3}
+                                    lineJoin="round"
+                                    lineCap="round"
+                                />
+                            ))}
+
+                            {this.props.map.getVertices().map((v) => (
+                                <RK.Group
+                                    key={v.id} 
+                                    onClick={(
+                                        event: KonvaEventObject<MouseEvent>
+                                    ) => {
+                                        {
+                                            this.select(v);
+                                            this.deselectCharacter(true);
+                                        }
+                                    }}
+                                >
+                                    <AgentImage
+                                        x={v.coords[0]}
+                                        y={v.coords[1]}
+                                        selected={this.props.selected === v}
+                                        data={v}
+                                        player={this.props.player}
+                                    />
+
+                                    {/* Debug text: */}
+
+                                    {/* <RK.Text
+                                        text={
+                                            v instanceof Agent
+                                                ? v.resources.toString()
+                                                : v instanceof DeadAgent
+                                                ? v.deadCount.toString()
+                                                : v.id.toString()
+                                        }
+                                        x={v.coords[0] - AGENT_RADIUS / 2}
+                                        y={v.coords[1] - AGENT_RADIUS / 2}
+                                        width={AGENT_RADIUS * 3}
+                                        height={AGENT_RADIUS * 3}
+                                        offsetX={AGENT_RADIUS * 3}
+                                        offsetY={AGENT_RADIUS}
+                                        fontSize={AGENT_RADIUS}
+                                        fill={v.isAlive() ? "black" : "red"}
+                                        shadowOffsetX={1}
+                                        shadowOffsetY={1}
+                                        align="center"
+                                        verticalAlign="middle"
+                                    /> */}
+                                </RK.Group>
+                            ))}
+                        </RK.Layer>
+
+                        <RK.Layer>
+                            {this.props.map.getVertices().map((v) => (
+                                <RK.Group key={v.id + " vertex-anim"}>
+                                    {v instanceof Agent && (
+                                        <AnimMood
+                                            turnCount={this.props.turnCount}
+                                            x={v.coords[0]}
+                                            y={v.coords[1]}
+                                            mood={
+                                                v instanceof Agent
+                                                    ? v.mood
+                                                    : undefined
+                                            }
+                                        />
+                                    )}
+                                    {v instanceof Agent && (
+                                        <AnimResources
+                                            turnCount={this.props.turnCount}
+                                            x={v.coords[0]}
+                                            y={v.coords[1]}
+                                            resources={
+                                                v instanceof Agent
+                                                    ? v.resources
+                                                    : undefined
+                                            }
+                                        />
+                                    )}
+                                    {v instanceof Agent && (
+                                        <AnimChangeIdeology
+                                            turnCount={this.props.turnCount}
+                                            x={v.coords[0]}
+                                            y={v.coords[1]}
+                                            ideology={
+                                                v instanceof Agent
+                                                    ? v.ideology
+                                                    : undefined
+                                            }
+                                        />
+                                    )}
+                                </RK.Group>
+                            ))}
+                        </RK.Layer>
+
+                        <RK.Layer>
+                            {this.props.map.getAllEdges().map(([v1, v2, e]) => (
+                                <RK.Group key={v1.id + " edge-anim " + v2.id}>
+                                    {v1 instanceof Agent &&
+                                        v2 instanceof Agent && (
+                                            <AnimChoice
+                                                turnCount={this.props.turnCount}
+                                                x1={v1.coords[0]}
+                                                y1={v1.coords[1]}
+                                                x2={v2.coords[0]}
+                                                y2={v2.coords[1]}
+                                                history={e.history}
+                                            />
+                                        )}
+                                    {v1 instanceof Agent &&
+                                        v2 instanceof Agent && (
+                                            <AnimInfluence
+                                                turnCount={this.props.turnCount}
+                                                x1={v1.coords[0]}
+                                                y1={v1.coords[1]}
+                                                x2={v2.coords[0]}
+                                                y2={v2.coords[1]}
+                                                influence={e.influence}
+                                            />
+                                        )}
+                                </RK.Group>
+                            ))}
+                        </RK.Layer>
+                    </RK.Stage>
+                </div>
+            </div>
+        );
+    }
+}
 /*
 function give() {
     alert("Gee thanks");
